+++ packages/react-magma-dom/src/components/DatePicker/index.tsx
@@ -118,7 +118,7 @@ const DatePickerCalendar = styled.div<{ opened: boolean }>`
   margin-top: -${props => props.theme.spaceScale.spacing07};
   opacity: ${props => (props.opened ? '1' : '0')};
   overflow: hidden;
-  position: absolute;
+  position: fixed;
   transition: opacity 0.2s ease-in-out 0s;
   width: 320px;
   z-index: ${props => (props.opened ? '998' : '-1')};
@@ -182,8 +182,7 @@ export const DatePicker = React.forwardRef<HTMLInputElement, DatePickerProps>(
       if (inDateRange(newDate, convertedMinDate, convertedMaxDate)) {
         return newDate;
       } else if (convertedMaxDate || convertedMinDate) {
-        return convertedMinDate? convertedMinDate
-          : convertedMaxDate;
+        return convertedMinDate ? convertedMinDate : convertedMaxDate;
       }
     }
 
@@ -293,11 +292,6 @@ export const DatePicker = React.forwardRef<HTMLInputElement, DatePickerProps>(
         event.preventDefault();
         setCalendarOpened(false);
         inputRef.current.focus();
-      }
-
-      if (event.key === '?') {
-        event.preventDefault();
-        setShowHelperInformation(true);
       }
     }