+++ packages/react-magma-dom/src/components/Drawer/index.ts
@@ -1,0 +1,2 @@
+// export * from "./useDrawer"
+export * from "./Drawer"