+++ packages/react-magma-dom/src/components/Table/Table.stories.tsx
@@ -11,8 +11,13 @@ import {
   TableProps,
   TableDensity,
   TableRowColor,
+  TableSortDirection,
+  TableCellAlign,
 } from './';
 import { magma } from '../../theme/magma';
+import { Announce } from '../Announce';
+import { VisuallyHidden } from '../VisuallyHidden';
+
 import { Story, Meta } from '@storybook/react/types-6-0';
 
 const rows = [
@@ -225,8 +230,8 @@ export const Pagination = () => {
   const [pageIndex, setPageIndex] = React.useState<number>(0);
   const [rowsPerPage, setRowsPerPage] = React.useState<number>(10);
 
-  function handleRowsPerPageChange(numberOfPages) {
-    setRowsPerPage(numberOfPages);
+  function handleRowsPerPageChange(numberOfRows) {
+    setRowsPerPage(numberOfRows);
     setPageIndex(0);
   }
 
@@ -420,5 +425,119 @@ export const RowColorsInverse = () => {
         </TableBody>
       </Table>
     </Card>
+  );
+};
+
+export const Sortable = () => {
+  const products = [
+    { id: 1, name: 'Cheese', price: 5, stock: 20 },
+    { id: 2, name: 'Milk', price: 5, stock: 32 },
+    { id: 3, name: 'Yogurt', price: 3, stock: 12 },
+    { id: 4, name: 'Heavy Cream', price: 10, stock: 9 },
+    { id: 5, name: 'Butter', price: 2, stock: 99 },
+    { id: 6, name: 'Sour Cream ', price: 5, stock: 86 },
+  ];
+
+  const [sortConfig, setSortConfig] = React.useState({
+    key: 'name',
+    direction: TableSortDirection.ascending,
+    message: '',
+  });
+
+  const sortedItems = React.useMemo(() => {
+    let sortableItems = [...products];
+    if (sortConfig !== null) {
+      sortableItems.sort((a, b) => {
+        if (a[sortConfig.key] < b[sortConfig.key]) {
+          return sortConfig.direction === TableSortDirection.ascending ? -1 : 1;
+        }
+        if (a[sortConfig.key] > b[sortConfig.key]) {
+          return sortConfig.direction === TableSortDirection.ascending ? 1 : -1;
+        }
+        return 0;
+      });
+    }
+    return sortableItems;
+  }, [products, sortConfig]);
+
+  const requestSort = key => {
+    let direction = TableSortDirection.ascending;
+    if (
+      sortConfig &&
+      sortConfig.key === key &&
+      sortConfig.direction === TableSortDirection.ascending
+    ) {
+      direction = TableSortDirection.descending;
+    }
+    const message = `Table is sorted by ${key}, ${direction}`;
+    setSortConfig({ key, direction, message });
+  };
+
+  return (
+    <>
+      <Card>
+        <Table>
+          <TableHead>
+            <TableRow>
+              <TableHeaderCell
+                onSort={() => {
+                  requestSort('name');
+                }}
+                isSortable
+                sortDirection={
+                  sortConfig.key === 'name'
+                    ? sortConfig.direction
+                    : TableSortDirection.none
+                }
+              >
+                Name
+              </TableHeaderCell>
+              <TableHeaderCell
+                onSort={() => {
+                  requestSort('price');
+                }}
+                isSortable
+                align={TableCellAlign.right}
+                sortDirection={
+                  sortConfig.key === 'price'
+                    ? sortConfig.direction
+                    : TableSortDirection.none
+                }
+              >
+                Price
+              </TableHeaderCell>
+              <TableHeaderCell
+                onSort={() => {
+                  requestSort('stock');
+                }}
+                isSortable
+                align={TableCellAlign.right}
+                sortDirection={
+                  sortConfig.key === 'stock'
+                    ? sortConfig.direction
+                    : TableSortDirection.none
+                }
+              >
+                In Stock
+              </TableHeaderCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {sortedItems.map(item => (
+              <TableRow key={item.id}>
+                <TableCell>{item.name}</TableCell>
+                <TableCell align={TableCellAlign.right}>
+                  ${item.price}
+                </TableCell>
+                <TableCell align={TableCellAlign.right}>{item.stock}</TableCell>
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </Card>
+      <Announce>
+        <VisuallyHidden>{sortConfig.message}</VisuallyHidden>
+      </Announce>
+    </>
   );
 };