+++ packages/react-magma-dom/src/index.ts
@@ -117,8 +117,10 @@ export {
   LoadingIndicator,
   LoadingIndicatorProps,
 } from './components/LoadingIndicator';
+export { NativeSelect, NativeSelectProps } from './components/NativeSelect';
 export { NavTabs } from './components/NavTabs';
 export { NavTab } from './components/NavTabs/NavTab';
+export { Pagination } from './components/Pagination';
 export { Paragraph, ParagraphProps } from './components/Paragraph';
 export { PasswordInput, PasswordInputProps } from './components/PasswordInput';
 export {
@@ -214,3 +216,7 @@ export { defaultI18n } from './i18n/default';
 export { I18nInterface } from './i18n/interface';
 export { useMediaQuery } from './hooks/useMediaQuery';
 export { InverseContext, useIsInverse } from './inverse';
+
+export * from './components/Accordion';
+export * from './components/Transition';
+export * from './components/Drawer';