+++ website/react-magma-docs/CHANGELOG.md
@@ -1,5 +1,70 @@
 # Change Log
 
+## 2.2.0-next.6
+
+### Patch Changes
+
+- Updated dependencies [1ceb9951]
+  - react-magma-dom@2.4.0-next.6
+  - @cengage-patterns/header@1.0.0-next.6
+
+## 2.2.0-next.5
+
+### Minor Changes
+
+- f0a1882f: fix/controlled-toggle
+
+  Update `Toggle` to have controlled and uncontrolled variants akin to `Checkbox`
+
+### Patch Changes
+
+- 6d2a2be6: chore: update icons
+- Updated dependencies [f0a1882f]
+- Updated dependencies [6d2a2be6]
+  - react-magma-dom@2.4.0-next.5
+  - @cengage-patterns/header@1.0.0-next.5
+
+## 2.1.12-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [88e2daee]
+  - react-magma-dom@2.4.0-alpha.4
+  - @cengage-patterns/header@1.0.0-alpha.4
+
+## 2.1.12-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [aaa4d559]
+  - react-magma-dom@2.4.0-alpha.3
+  - @cengage-patterns/header@1.0.0-alpha.3
+
+## 2.1.12-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [b7f7e431]
+- Updated dependencies [c49cda0e]
+  - react-magma-dom@2.4.0-alpha.2
+  - @cengage-patterns/header@1.0.0-alpha.2
+
+## 2.1.12-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [efb6f297]
+  - react-magma-dom@2.3.13-alpha.1
+  - @cengage-patterns/header@0.1.3-alpha.1
+
+## 2.1.12-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [57e28ac7]
+  - react-magma-dom@2.3.13-alpha.0
+  - @cengage-patterns/header@0.1.3-alpha.0
+
 ## 2.1.11
 
 ### Patch Changes